Return quasi-continuation when validating items constraints
Rather than validating a jerk_constraint:items/2 constraint directly
for all elements in an array we break down the constraint into a list
of simpler validations (type + constraints) that must be satisfied.

This lets us keep the type-checking and general object validation
outside of the constraint module and keeps the constraint module
focused on single-constraint validation.

The continuation structure will need to change some to accomodate the
`contains` constraint which can set limits on the number of items in
an array that satisfy the constraint, but I'll cross that bridge when
I get to it.