Handle continuations adjacent to valid primitive props in validator
There was a bug that appeared when an objects properties had valid
values with primitive (or non-reference) types along side reference
types that need to return a continuation for further
validation. Because we need to handle a mix of boolean and `{continue,
_}` return values in the check_params fold fun we cannot just do
simple boolean logic like it was before. Instead we will do some
pattern matching to break things down and put them back together
nicely