Add suppport for arrays containing jerk objects
Handling these is tricky because when we return the array we need to
tag the objects it contains with the schema so they can be used as
jerk terms; however, the validation (and the rest of jerk) assumes
that only the top-level object is tagged with its schema so we must
traverse the array and add the schema tag to its elements when it is
returned.

TODO still need to test arrays with mixed type elements (e.g. objects
and ints) and arrays with unconstrained types.