Return the correct schema id for nested references in an array
This fixes a subtle bug that only seems to arrise for - bear with me -
schemas with arrays that contain objects constrained by reference that
contain arrays of objects constrained by reference. At the second
level the schema ID got clobbered and became invalid. Then later if
you tried to use `jerk:is_object/1` on the objects from the innermost
array it would incorrectly return false - because the schema couldn't
be resolved. There would probably also be a problem if you tried to
update a property in those objects.